Move check into devmode
