+++ db/change_listener.go
@@ -14,9 +14,7 @@ import (
 	"bytes"
 	"context"
 	"expvar"
-	"fmt"
 	"math"
-	"os"
 	"sync"
 	"time"
 
@@ -537,19 +535,4 @@ func (waiter *ChangeWaiter) RefreshUserKeys(user auth.User, metaKeys *base.Metad
 func (db *Database) NewUserWaiter() *ChangeWaiter {
 	trackUnusedSequences := false
 	return db.mutationListener.NewWaiterWithChannels(channels.Set{}, db.User(), trackUnusedSequences)
-}
-
-// GetCachingFeedDelay returns the delay to apply to processing of each DCP event in the caching feed, used for testing
-// purposes to simulate a slow feed.
-func GetCachingFeedDelay() (time.Duration, error) {
-	delayEnvVar := "SG_TEST_CACHING_FEED_DELAY"
-	d := os.Getenv(delayEnvVar)
-	if d == "" {
-		return 0, nil
-	}
-	delay, err := time.ParseDuration(d)
-	if err != nil {
-		return 0, fmt.Errorf("setting %s=%s is not a valid time: %w", delayEnvVar, d, err)
-	}
-	return delay, nil
 }
